Implemented the Update command

+++ src/logic/Logic.java
@@ -2,21 +2,17 @@ package logic;
 
 import parser.MainParser;
 import storage.Storage;
-import ui.MessageObserver;
 
-import java.text.ParseException;
 import java.util.ArrayList;
-import java.util.Collections;
-import java.util.Comparator;
 import java.util.List;
 import java.util.Observable;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
 import models.Commands;
-import models.DeadlineTask;
+import models.Deadline;
 import models.Event;
-import models.FloatingTask;
+import models.Todo;
 import models.ParsedObject;
 import models.Task;
 
@@ -44,16 +40,20 @@ public class Logic extends Observable {
 				break;
 			case DISPLAY:
 				//display(input);
+				break;
 			case SEARCH:
 				//search(input);
+				break;
 			case UPDATE:
-				//update(input);
+				update(input);
+				break;
 			case DELETE:
 				//return delete(input);
 				delete(input);
 				break;
 			case UNDO:
 				//undo(input);
+				break;
 			default:
 				observable.updateStatusMsg("Invalid command entered. Please try again.");
 		}
@@ -89,24 +89,24 @@ public class Logic extends Observable {
 					observable.updateTables(Commands.TASK_TYPE.EVENT);
 					break;
 					//return "<html><b>\"" + dEvt.getTaskDesc() + "\"</b><br/>has been successfully added as an Event from <b>" + parser.formatDate(dEvt.getFromDate(), "EEE, d MMM yyyy h:mm a") + "</b> to <b>" + parser.formatDate(dEvt.getToDate(), "EEE, d MMM yyyy h:mm a") + "</b>.</html>";
-				case FLOATING_TASK:
-					FloatingTask flt = (FloatingTask)v.get(i);
+				case TODO:
+					Todo flt = (Todo)v.get(i);
 					logger.log(Level.FINE, flt.getTaskID() + ", " + flt.getTaskDesc());
 					//System.out.println(flt.getTaskID() + ", " + flt.getTaskDesc());
-					Storage.addTask(flt, Commands.TASK_TYPE.FLOATING_TASK);
-					Storage.saveTaskType(Commands.TASK_TYPE.FLOATING_TASK);
+					Storage.addTask(flt, Commands.TASK_TYPE.TODO);
+					Storage.saveTaskType(Commands.TASK_TYPE.TODO);
 					observable.updateStatusMsg("<html><b>\"" + flt.getTaskDesc() + "\"</b><br/>has been successfully added as a Todo task.</html>");
-					observable.updateTables(Commands.TASK_TYPE.FLOATING_TASK);
+					observable.updateTables(Commands.TASK_TYPE.TODO);
 					break;
 					//return "<html><b>\"" + flt.getTaskDesc() + "\"</b><br/>has been successfully added as a Todo task.</html>";
-				case DEADLINE_TASK:
-					DeadlineTask dt = (DeadlineTask)v.get(i);
+				case DEADLINE:
+					Deadline dt = (Deadline)v.get(i);
 					logger.log(Level.FINE, dt.getTaskID() + ", " + dt.getTaskDesc() + ", " + dt.getDate());
 					//System.out.println(dt.getTaskID() + ", " + dt.getTaskDesc() + ", " + dt.getDate());
-					Storage.addTask(dt, Commands.TASK_TYPE.DEADLINE_TASK);
-					Storage.saveTaskType(Commands.TASK_TYPE.DEADLINE_TASK);
+					Storage.addTask(dt, Commands.TASK_TYPE.DEADLINE);
+					Storage.saveTaskType(Commands.TASK_TYPE.DEADLINE);
 					observable.updateStatusMsg("<html><b>\"" + dt.getTaskDesc() + "\"</b><br/>has been successfully added as a Deadline task that must be completed by <b>" + parser.formatDate(dt.getDate(), "EEE, d MMM yyyy") + "</b>.</html>");
-					observable.updateTables(Commands.TASK_TYPE.DEADLINE_TASK);
+					observable.updateTables(Commands.TASK_TYPE.DEADLINE);
 					break;
 					//return "<html><b>\"" + dt.getTaskDesc() + "\"</b><br/>has been successfully added as a Deadline task that must be completed by <b>" + parser.formatDate(dt.getDate(), "EEE, d MMM yyyy") + "</b>.</html>";
 				default:
@@ -126,19 +126,24 @@ public class Logic extends Observable {
 		System.out.println(obj.getCommandType());
 		System.out.println(obj.getTaskType());
 		ArrayList<Integer> v = obj.getObjects();
-		String statusMsg = "<html>Tasks IDs: <b>";
-		for (int i = 0; i < v.size(); i++) {
-			Storage.delete(v.get(i));
-			statusMsg += v.get(i);
-			System.out.print(v.get(i));
-			if (i < v.size()-1) {
-				statusMsg += ", ";
-				System.out.print("|");
+		String statusMsg = "<html>";
+		if (v.size() > 0) {
+			statusMsg += "Tasks IDs: ";
+			for (int i = 0; i < v.size(); i++) {
+				Storage.delete(v.get(i));
+				statusMsg += ("<b>" + v.get(i) + "</b>");
+				System.out.print(v.get(i));
+				if (i < v.size()-1) {
+					statusMsg += ", ";
+					System.out.print("|");
+				}
 			}
+			Storage.saveAllTask();
+			statusMsg += "<br/>have been deleted successfully.</html>";
+		} else {
+			statusMsg += "TaskIDs are not found.</html>";
 		}
 		System.out.println();
-		statusMsg += "</b> have been deleted successfully.<br/></html>";
-		Storage.saveAllTask();
 		observable.updateStatusMsg(statusMsg);
 		observable.updateTables(Commands.TASK_TYPE.ALL);
 		// return Storage.delete(parser.getDeleteParsedObject(input));
@@ -167,11 +172,12 @@ public class Logic extends Observable {
 		// return Storage.display(parser.getDisplayParsedObject(input));
 	}
 
+	/*
 	public ArrayList<List<Task>> getAllTaskLists() {
 		ArrayList<List<Task>> taskLists = new ArrayList<List<Task>>();
 		taskLists.add(Storage.getAllTask(Commands.TASK_TYPE.EVENT));
-		taskLists.add(Storage.getAllTask(Commands.TASK_TYPE.FLOATING_TASK));
-		taskLists.add(Storage.getAllTask(Commands.TASK_TYPE.DEADLINE_TASK));
+		taskLists.add(Storage.getAllTask(Commands.TASK_TYPE.TODO));
+		taskLists.add(Storage.getAllTask(Commands.TASK_TYPE.DEADLINE));
 		return taskLists;
 	}
 
@@ -179,12 +185,12 @@ public class Logic extends Observable {
 		return (Storage.getAllTask(Commands.TASK_TYPE.EVENT).size() != 0) ? (ArrayList)Storage.getAllTask(Commands.TASK_TYPE.EVENT) : new ArrayList();
 	}
 
-	public ArrayList<FloatingTask> getAllFloatingTasks() {
-		return (Storage.getAllTask(Commands.TASK_TYPE.FLOATING_TASK).size() != 0) ? (ArrayList)Storage.getAllTask(Commands.TASK_TYPE.FLOATING_TASK) : new ArrayList();
+	public ArrayList<Todo> getAllTodos() {
+		return (Storage.getAllTask(Commands.TASK_TYPE.TODO).size() != 0) ? (ArrayList)Storage.getAllTask(Commands.TASK_TYPE.TODO) : new ArrayList();
 	}
 
-	public ArrayList<DeadlineTask> getAllDeadlineTasks() {
-		return (Storage.getAllTask(Commands.TASK_TYPE.DEADLINE_TASK).size() != 0) ? (ArrayList)Storage.getAllTask(Commands.TASK_TYPE.DEADLINE_TASK) : new ArrayList();
+	public ArrayList<Deadline> getAllDeadlineTasks() {
+		return (Storage.getAllTask(Commands.TASK_TYPE.DEADLINE).size() != 0) ? (ArrayList)Storage.getAllTask(Commands.TASK_TYPE.DEADLINE) : new ArrayList();
 	}
 
 	private ArrayList<Event> searchEvents(String input) {
@@ -201,11 +207,11 @@ public class Logic extends Observable {
 		return matches;
 	}
 
-	private ArrayList<FloatingTask> searchFloatingTasks(String input) {
+	private ArrayList<Todo> searchTodos(String input) {
 		ParsedObject obj = parser.getSearchParsedObject(input);
-		ArrayList<FloatingTask> matches = new ArrayList<FloatingTask>();
+		ArrayList<Todo> matches = new ArrayList<Todo>();
 
-		for (FloatingTask ft : getAllFloatingTasks()) {
+		for (Todo ft : getAllTodos()) {
 			for (String s : (ArrayList<String>)obj.getObjects()) {
 				if (ft.getTaskDesc().toLowerCase().contains(s)) {
 					matches.add(ft);
@@ -215,11 +221,11 @@ public class Logic extends Observable {
 		return matches;
 	}
 
-	private ArrayList<DeadlineTask> searchDeadlineTasks(String input) {
+	private ArrayList<Deadline> searchDeadlineTasks(String input) {
 		ParsedObject obj = parser.getSearchParsedObject(input);
-		ArrayList<DeadlineTask> matches = new ArrayList<DeadlineTask>();
+		ArrayList<Deadline> matches = new ArrayList<Deadline>();
 
-		for (DeadlineTask dt : getAllDeadlineTasks()) {
+		for (Deadline dt : getAllDeadlineTasks()) {
 			for (String s : (ArrayList<String>)obj.getObjects()) {
 				if (dt.getTaskDesc().toLowerCase().contains(s)) {
 					matches.add(dt);
@@ -244,5 +250,24 @@ public class Logic extends Observable {
 		System.out.println();
 		return null;
 		// return Storage.search(parser.getSearchParsedObject(input));
+	}*/
+
+	private void update(String input) {
+		ParsedObject obj = parser.getUpdateParsedObject(input);
+		ArrayList<String> params = obj.getObjects();
+		Task t = Storage.getTaskByID(parser.parseInteger(params.get(0)));
+		if (t != null) {
+			if (Update.update(obj)) {
+				observable.updateStatusMsg("Successfully updated.");
+				observable.updateTables(Commands.TASK_TYPE.ALL);
+				Storage.saveAllTask();
+			} else {
+				observable.updateStatusMsg("Invalid column or value entered.");
+			}
+		}
+	}
+
+	public Task getTaskByID(int id) {
+		return Storage.getTaskByID(id);
 	}
 }
